requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
aws==0.2.5
awscrt==0.29.2
bcrypt==5.0.0
boto==2.49.0
boto3==1.42.39
botocore==1.42.39
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.3.1
distro==1.9.0
fabric==3.2.2
frozenlist==1.8.0
groq==0.36.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
invoke==2.2.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jiter==0.12.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.7.0
langchain-core==1.1.0
langchain-groq==1.0.1
# Removed incompatible legacy package pins:
# langchain==0.1.20
# langchain-openai==0.1.6
# Keeping langchain-core and modern integrations (langchain-core, langchain-perplexity)
langchain-perplexity==1.1.0
# langchain-text-splitters pinned version removed to avoid conflicts with langchain-core
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.46
marshmallow==3.26.1
more-itertools==10.8.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==2.16.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==23.2
paramiko==4.0.0
prettytable==3.17.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.5.3
wrapt==2.1.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
